Move and test analyze_code
